shallot: rename recipe to add
